project/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.12",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "comm==0.2.2",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "executing==2.2.0",
    "fastapi>=0.115.13",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.7.0",
    "fsspec==2025.5.1",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-search-results==2.4.2",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "grpcio==1.73.0",
    "grpcio-status==1.73.0",
    "h11==0.16.0",
    "hf-xet==1.1.3",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.33.0",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "langchain==0.3.25",
    "langchain-community==0.3.25",
    "langchain-core==0.3.65",
    "langchain-google-genai==2.1.5",
    "langchain-huggingface==0.3.0",
    "langchain-mistralai>=0.2.10",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.45",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.3.0",
    "orjson==3.10.18",
    "packaging==24.2",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pymongo>=4.13.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "ruff>=0.12.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "tokenizers==0.21.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "uvicorn>=0.34.3",
    "wcwidth==0.2.13",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
